Compilation fix
git-svn-id: file:///home/roberto/Git/svn@12421 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 572ab355c5c9930029e99950a60675c93ac369fc

+++ trunk/xmipp/libraries/data/multidim_array.h
@@ -2382,13 +2382,13 @@ public:
         y -= STARTINGY(*this);
         x -= STARTINGX(*this);
 
-        int l1 = ceil(x - SplineDegree_1);
+        int l1 = (int)ceil(x - SplineDegree_1);
         int l2 = l1 + SplineDegree;
 
-        int m1 = ceil(y - SplineDegree_1);
+        int m1 = (int)ceil(y - SplineDegree_1);
         int m2 = m1 + SplineDegree;
 
-        int n1 = ceil(z - SplineDegree_1);
+        int n1 = (int)ceil(z - SplineDegree_1);
         int n2 = n1 + SplineDegree;
 
         double zyxsum = 0.0;
